moved servodata.h and servoports.h to /config and using them in mp-robot-d.h as well
