+++ config/mp-robot-d.h
@@ -1,4 +1,4 @@
-#include "Arduino.h"
+#include "servodata.h"
 
 // ---- Servos ---- on Teensy PWM pins are 3-6 9-10 20-21
 #define pin_servo_0 3
